+++ speckle-powerbi.code-workspace
@@ -1,17 +1,17 @@
 {
   "folders": [
     {
-      "name": "root",
-      "path": "."
+      "name": "👀 powerbi-visual",
+      "path": "src/powerbi-visual"
     },
     {
-      "name": "DataConnector",
+      "name": "➡️ powerbi-data-connector",
       "path": "src/powerbi-data-connector"
     },
     {
-      "name": "Visual",
-      "path": "src/powerbi-visual"
-    }
+      "name": "🏠 root",
+      "path": "."
+    },
   ],
   "settings": {
     "powerquery.general.mode": "SDK",
@@ -23,24 +23,30 @@
       "**/node_modules/**": true,
       ".tmp": true
     },
-    "files.exclude": {
-      ".tmp": true
-    },
     "search.exclude": {
       ".tmp": true,
-      "typings": true
+      "typings": true,
+      "dist": true,
+      "wepbpack.statistics.dev.html": true,
+      "wepbpack.statistics.html": true,
     },
     "json.schemas": [
       {
-        "fileMatch": ["/pbiviz.json"],
+        "fileMatch": [
+          "/pbiviz.json"
+        ],
         "url": "./src/powerbi-visual/node_modules/powerbi-visuals-api/schema.pbiviz.json"
       },
       {
-        "fileMatch": ["/capabilities.json"],
+        "fileMatch": [
+          "/capabilities.json"
+        ],
         "url": "./src/powerbi-visual/node_modules/powerbi-visuals-api/schema.capabilities.json"
       },
       {
-        "fileMatch": ["/dependencies.json"],
+        "fileMatch": [
+          "/dependencies.json"
+        ],
         "url": "./src/powerbi-visual/node_modules/powerbi-visuals-api/schema.dependencies.json"
       }
     ]
