+++ src/powerbi-visual/src/App.vue
@@ -1,13 +1,17 @@
 <script setup lang="ts">
 import HomeView from './views/HomeView.vue'
 import ViewerView from './views/ViewerView.vue'
-import { computed } from 'vue'
+import { computed, onMounted } from 'vue'
 import { useStore } from 'vuex'
 import { storeKey } from 'src/injectionKeys'
 
 let store = useStore(storeKey)
 let status = computed(() => {
   return store.state.status
+})
+
+onMounted(() => {
+  console.log("App mounted")
 })
 </script>
